refactor: extract useFormField hook and related context into separate file
